update them owner moi duoc dang bai

+++ backend/rental_post/api/views.py
@@ -8,6 +8,7 @@ from rest_framework.permissions import IsAuthenticated
 from rest_framework.exceptions import NotFound
 from rest_framework.pagination import PageNumberPagination
 from django.core.cache import cache
+from backend import settings
 
 # API xem toàn bộ bài đăng, tạo bài đăng của người dùng hiện tại
 class RentalPostListCreateAPIView(GenericAPIView):
@@ -20,16 +21,27 @@ class RentalPostListCreateAPIView(GenericAPIView):
 
     def get(self, request):
         """Lấy danh sách bài đăng với phân trang"""
-        page_number = request.query_params.get('page', 1)
+        if not request.user.is_authenticated:
+            return Response({
+                "success": False,
+                "message": "Bạn chưa đăng nhập."
+            }, status=status.HTTP_401_UNAUTHORIZED)
 
-        # Cache
-        cached_posts = cache.get(f'rental_posts_cache_page_{page_number}_{request.user.id}')
+        if request.user.role != 'owner':
+            return Response({
+                "success": False,
+                "message": "Bạn chưa là Owner để truy cập vào danh sách bài đăng."
+            }, status=status.HTTP_403_FORBIDDEN)
+
+        page_number = request.query_params.get('page', 1)
+        cache_key = f'rental_posts_cache_page_{page_number}_{request.user.id}'
+        cached_posts = cache.get(cache_key)
+
         if cached_posts:
             cached_posts["success"] = True
             return Response(cached_posts, status=status.HTTP_200_OK)
 
-        posts = self.get_queryset()  # dùng lại hàm get_queryset()
-
+        posts = self.get_queryset()
         paginator = self.pagination_class()
         result_page = paginator.paginate_queryset(posts, request)
         serializer = self.get_serializer(result_page, many=True)
@@ -44,30 +56,44 @@ class RentalPostListCreateAPIView(GenericAPIView):
                 "data": serializer.data
             }
         }
-        cache.set(f'rental_posts_cache_page_{page_number}_{request.user.id}', result_data)
 
+        cache.set(cache_key, result_data)
         return Response(result_data, status=status.HTTP_200_OK)
+
 
 
     def post(self, request):
         """Tạo bài đăng mới"""
+        if request.user.role != 'owner':
+            return Response({
+                "success": False,
+                "message": "Bạn phải là Owner để tạo bài đăng."
+            }, status=status.HTTP_403_FORBIDDEN)
+        
         serializer = RentalPostSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save(user=request.user)
 
-            cache.delete('rental_posts_cache')
+            total_items = RentalPost.objects.filter(user=request.user).count()
+            page_size = settings.REST_FRAMEWORK.get('PAGE_SIZE', 6)
+            total_pages = (total_items + page_size - 1) // page_size
+
+            for page in range(1, total_pages + 1):
+                cache_key = f'rental_posts_cache_page_{page}_{request.user.id}'
+                cache.delete(cache_key)
 
             return Response({
                 "success": True,
                 "message": "Bài đăng đã được tạo thành công.",
                 "data": serializer.data
             }, status=status.HTTP_201_CREATED)
-
+        
         return Response({
             "success": False,
             "message": "Tạo bài đăng thất bại.",
             "errors": serializer.errors
         }, status=status.HTTP_400_BAD_REQUEST)
+
 
 # API xem chi tiết bài đăng, sửa bài đăng, xóa bài đăng theo ID của người dùng hiện tại
 class RentalPostDetailUpdateDeleteAPIView(APIView):
@@ -78,7 +104,7 @@ class RentalPostDetailUpdateDeleteAPIView(APIView):
         try:
             post = RentalPost.objects.get(id=id, user=request.user)
         except RentalPost.DoesNotExist:
-            raise NotFound(detail="Bài đăng không tồn tại.")
+            raise NotFound(detail="Không tìm thấy bài đăng của bạn.")
 
         serializer = RentalPostSerializer(post)
         return Response({
@@ -89,14 +115,27 @@ class RentalPostDetailUpdateDeleteAPIView(APIView):
 
     def put(self, request, id):
         """Sửa bài đăng (cập nhật bài đăng theo ID)"""
+        if request.user.role != 'owner':
+            return Response({
+                "success": False,
+                "message": "Bạn phải là Owner để sửa bài đăng."
+            }, status=status.HTTP_403_FORBIDDEN)
+            
         try:
             post = RentalPost.objects.get(id=id, user=request.user)
         except RentalPost.DoesNotExist:
-            raise NotFound(detail="Bài đăng không tồn tại.")
+            raise NotFound(detail="Không tìm thấy bài đăng của bạn.")
 
         serializer = RentalPostSerializer(post, data=request.data, partial=False)
         if serializer.is_valid():
             serializer.save()
+            total_items = RentalPost.objects.filter(user=request.user).count()
+            page_size = settings.REST_FRAMEWORK.get('PAGE_SIZE', 6)
+            total_pages = (total_items + page_size - 1) // page_size
+
+            for page in range(1, total_pages + 1):
+                cache_key = f'rental_posts_cache_page_{page}_{request.user.id}'
+                cache.delete(cache_key)
             return Response({
                 "success": True,
                 "message": "Cập nhật bài đăng thành công.",
@@ -111,11 +150,23 @@ class RentalPostDetailUpdateDeleteAPIView(APIView):
 
     def delete(self, request, id):
         """Xóa bài đăng"""
+        if request.user.role != 'owner':
+            return Response({
+                "success": False,
+                "message": "Bạn phải là Owner để xóa bài đăng."
+            }, status=status.HTTP_403_FORBIDDEN)
         try:
             post = RentalPost.objects.get(id=id, user=request.user)
         except RentalPost.DoesNotExist:
-            raise NotFound(detail="Bài đăng không tồn tại.")
+            raise NotFound(detail="Không tìm thấy bài đăng của bạn.")
         post.delete()
+        total_items = RentalPost.objects.filter(user=request.user).count()
+        page_size = settings.REST_FRAMEWORK.get('PAGE_SIZE', 6)
+        total_pages = (total_items + page_size - 1) // page_size
+
+        for page in range(1, total_pages + 1):
+            cache_key = f'rental_posts_cache_page_{page}_{request.user.id}'
+            cache.delete(cache_key)
         return Response({
             "success": True,
             "message": "Bài đăng đã được xóa."
